Add the 16-colour block families (wool, concrete, terracotta, glass)
Every Minecraft dye colour as first-class blocks so imported builds keep their
palette instead of collapsing to a single grey stand-in. All tiles are generated
procedurally by tinting one base pattern per family — no hand art.

- js/core/colors.js — the 16 canonical dye colours (id → hex), shared by the
  block registry and the texture atlas.
- textures.js tints six patterns per colour: woven wool, flat concrete, grainy
  concrete powder, earthy terracotta, glazed terracotta (diagonal accent) and
  translucent stained glass. Atlas grown to 22 rows for the ~96 new tiles.
- blocks.js registers eight blocks per colour — wool, carpet (reuses the wool
  tile), concrete, concrete powder, terracotta, glazed terracotta, stained glass
  (transparent) and stained-glass pane — plus plain terracotta. 128 new blocks.
- Chunk block storage widened Uint8Array → Uint16Array (the only 8-bit id store;
  saves persist edits as plain numbers, so no save-format change) since the id
  count now exceeds 256.

Screenshot-verified: a 16×6 grid shows every colour reading distinctly across all
six materials, with glass translucent and glazed patterned. tests/unit/colors.test.mjs
(4 tests) + tests/_colorshot.mjs. 91 unit tests + smoke green.

+++ js/world/world.js
@@ -60,7 +60,7 @@ export class World {
   getChunk(cx, cz) { return this.chunks.get(chunkKey(cx, cz)); }
 
   generateChunk(cx, cz) {
-    const blocks = new Uint8Array(CHUNK * CHUNK * WORLD_H);
+    const blocks = new Uint16Array(CHUNK * CHUNK * WORLD_H); // 16-bit: >256 block ids (colored families, shapes)
     const chunk = { cx, cz, blocks, nodes: [], spawns: [], surfaceH: new Int16Array(CHUNK * CHUNK) };
     const setLocal = (lx, y, lz, id) => { blocks[lidx(lx, y, lz)] = id; };
     const gen = this.gen;
